# src/frontend/sqeez/src/api/api.yaml
openapi: 3.1.1
info:
  title: Sqeez.Api | v1
  version: 1.0.0
servers:
  - url: http://localhost:5000/
paths:
  /api/admins:
    get:
      tags:
        - Admins
      parameters:
        - name: PhoneNumber
          in: query
          schema:
            type: string
        - name: StrictRoleOnly
          in: query
          schema:
            type: boolean
        - name: Department
          in: query
          schema:
            type: string
        - name: SearchTerm
          in: query
          schema:
            type: string
        - name: IsOnline
          in: query
          schema:
            type: boolean
        - name: SchoolClassId
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
        - name: IsArchived
          in: query
          schema:
            type: boolean
        - name: PageNumber
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
        - name: PageSize
          in: query
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int32
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/PagedResponseOfAdminDto"
            application/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfAdminDto"
            text/json:
              schema:
                $ref: "#/components/schemas/PagedResponseOfAdminDto"
  /api/admins/{id}:
    get:
      tags:
        - Admins
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/AdminDto"
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDto"
            text/json:
              schema:
                $ref: "#/components/schemas/AdminDto"
    patch:
      tags:
        - Admins
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PatchAdminDto"
          text/json:
            schema:
              $ref: "#/components/schemas/PatchAdminDto"
          application/*+json:
            schema:
              $ref: "#/components/schemas/PatchAdminDto"
        required: true
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                $ref: "#/components/schemas/AdminDto"
            application/json:
              schema:
                $ref: "#/components/schemas/AdminDto"
            text/json:
              schema:
                $ref: "#/components/schemas/AdminDto"
    delete:
      tags:
        - Admins
      parameters:
        - name: id
          in: path
          required: true
          schema:
            pattern: ^-?(?:0|[1-9]\d*)$
            type:
              - integer
              - string
            format: int64
      responses:
        "200":
          description: OK
          content:
            text/plain:
              schema:
                type: boolean
            application/json:
              schema:
                type: boolean
            text/json:
              schema:
                type: boolean
  /api/auth/register:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
          text/json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
          application/*+json:
            schema:
              $ref: "#/components/schemas/RegisterDTO"
        required: true
      responses:
        "200":
          description: OK
  /api/auth/login:
    post:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginDTO"
          text/json:
            schema:
              $ref: "#/components/schemas/LoginDTO"
          application/*+json:
            schema:
              $ref: "#/components/schemas/LoginDTO"
        required: true
      responses:
        "200":
          description: OK
...
  /api/auth/elevate:
    patch:
      tags:
        - Auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleDTO"
          text/json:
            schema:
              $ref: "#/components/schemas/UpdateRoleDTO"
          application/*+json:
            schema:
              $ref: "#/components/schemas/UpdateRoleDTO"
        required: true
      responses:
        "200":
          description: OK
...
components:
  schemas:
    AdminDto:
      type: object
      properties:
        phoneNumber:
          type: string
        department:
          type:
            - "null"
            - string
        managedClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        username:
          type: string
        email:
          type: string
        currentXP:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        role:
          type: string
        avatarUrl:
          type:
            - "null"
            - string
        lastSeen:
          type: string
          format: date-time
        schoolClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
...
    LoginDTO:
      required:
        - email
        - password
      type: object
      properties:
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
...
    PagedResponseOfAdminDto:
      type: object
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/AdminDto"
        totalCount:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageNumber:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        pageSize:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
        totalPages:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int32
...
    PatchAdminDto:
      type: object
      properties:
        phoneNumber:
          type:
            - "null"
            - string
        department:
          type:
            - "null"
            - string
        managedClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        username:
          type:
            - "null"
            - string
        email:
          type:
            - "null"
            - string
        schoolClassId:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - "null"
            - integer
            - string
          format: int64
        avatarUrl:
          type:
            - "null"
            - string
...
    RegisterDTO:
      required:
        - username
        - email
        - password
      type: object
      properties:
        username:
          type: string
        email:
          type: string
        password:
          type: string
        rememberMe:
          type: boolean
          default: false
...
    UpdateRoleDTO:
      required:
        - id
        - role
      type: object
      properties:
        id:
          pattern: ^-?(?:0|[1-9]\d*)$
          type:
            - integer
            - string
          format: int64
        role:
          $ref: "#/components/schemas/UserRole"
        department:
          type:
            - "null"
            - string
        phoneNumber:
          type:
            - "null"
            - string
...
    UserRole:
      enum:
        - Student
        - Teacher
        - Admin
...